requirements.txt
azure-cognitiveservices-speech==1.38.0
python-dotenv==1.0.1
torch>=2.9.0
torchaudio>=2.9.0
numpy>=2.0.0
sounddevice>=0.5.0
soundfile>=0.12.1    
silero-vad>=6.0.0
requests>=2.31.0       
